requirements.txt
PySide6>=6.11.0
PyYAML>=6.0.3
Nuitka>=4.0.7
